Convert common game utilities to JavaScript
