+++ components/apps/Games/common/index.ts
@@ -1,5 +1,5 @@
-export { default as Overlay } from './Overlay';
-export { default as VirtualPad } from './VirtualPad';
+export { default as Overlay } from './Overlay.js';
+export { default as VirtualPad } from './VirtualPad.js';
 export { default as useGamepad } from './useGamepad';
 export { default as useGameLoop } from './useGameLoop';
 export { default as useSaveSlots } from './useSaveSlots';
